# Splitting on Lat/Lon
We can think of our solution as a series of transformations. The first
transformation we need to do is splitting our comma separated lat/lon
string into two separate strings. One that holds the latitude of our
coordinate, and the other that holds the longitude.

    expect(splitOnLatLon("48° 12′ 30″ N, 16° 22′ 23″ E")).to.deep.equal([
        "48° 12′ 30″ N",
        "16° 22′ 23″ E"
    ]);

We can test that a function called `splitOnLatLon` does just this.

Implementing `splitOnLatLon` is just a matter of stringing together a
few [Lodash](https://lodash.com/) function calls.