# The Haversine Formula
The next step in our transformation is using the two sets of latidudes
and longitudes we've constructured to calculate the distance between
these two points.

We would expect the same two points to have zero distance between them:

    expect(haversine(
        48.2083, 16.3730,
        48.2083, 16.3730,
        6371
    )).to.be.closeTo(0, 0.001);

And we would expect another set of points to have a resonable amount of
distance between them:

    expect(haversine(
        48.2083, 16.3730,
        16.3730, 48.2083,
        6371
    )).to.be.closeTo(3133.445, 0.001);

The `haversize` function is a fairly uninteresting implementation of the
[Haversine Formula](http://andrew.hedges.name/experiments/haversine/).
After implementing this formula, our tests pass!