# Finishing the Transformation
Now that we have all of the pieces of our transformation we can refactor
our `distance` function.

The basic idea is that we want to split our the lat/lon of each
coordinate, convert those DMS coordinates to a decimal format, pass the
resulting coordinates into `haversine` and then rounding the result.

After doing this refactoring, our tests still pass!